pyproject.toml
[project]
name = "retention-dashboard"
version = "1.0.0"
description = "Retention Metrics Dashboard with Flask backend"
requires-python = ">=3.9,<3.13"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pandas>=2.1.4",
    "numpy>=1.26.2,<2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

